requirements.txt
python-nmap
rich
requests
